+++ src/renderer/components/animator/Animator/Animator.tsx
@@ -10,14 +10,25 @@ import { Timeline } from "../Timeline/Timeline";
 import TitleToolbar from "../TitleToolbar/TitleToolbar";
 import "./Animator.css";
 import { LeftPanes } from "../LeftPanes/LeftPanes";
+import { useEffect } from "react";
+import { displayProjectTitle } from "../../../services/project/projectBuilder";
+import { useSelector } from "react-redux";
+import { RootState } from "../../../redux/store";
 
 export const Animator = () => {
+  const project = useSelector((state: RootState) => state.project.project);
+  useEffect(() => {
+    document.title = project
+      ? `${displayProjectTitle(project)} - Boats Animator`
+      : "Boats Animator";
+  }, [project]);
+
   return (
     <Page>
       <TitleToolbar />
       <PageBody>
         <Content>
-          <Group className="animator__overlay-tabs-pane-and-preview" gap="xs">
+          <div className="animator__overlay-tabs-pane-and-preview">
             <LeftPanes />
 
             <Stack flex={1} gap="xs">
@@ -26,7 +37,7 @@ export const Animator = () => {
             </Stack>
 
             <CapturePane />
-          </Group>
+          </div>
 
           <FrameToolbar />
           <Timeline />
