+++ src/renderer/components/common/App/App.tsx
@@ -1,34 +1,19 @@
-import { Navigate, Outlet, Route, RouterProvider, Routes } from "react-router-dom";
-import { PageRoute } from "../../../../common/PageRoute";
+import { Outlet } from "react-router-dom";
 import { FileManagerContextProvider } from "../../../context/FileManagerContext/FileManagerContextProvider";
 import { PersistedDirectoriesContextProvider } from "../../../context/PersistedDirectoriesContext/PersistedDirectoriesContextProvider";
-import { Theme } from "../../ui/Theme/Theme";
-import AppListener from "../AppListener/AppListener";
-import { useAnimatorRoutesAndProviders } from "./useAnimatorRoutesAndProviders";
-import { useStartupRoutes } from "./useStartupRoutes";
 import { AppErrorBoundary } from "../AppErrorBoundary/AppErrorBoundary";
-import { router } from "../../../router";
+import AppListener from "../AppListener/AppListener";
 
-const App = (): JSX.Element => {
-  // const startupRoutes = useStartupRoutes();
-  // const animatorRoutes = useAnimatorRoutesAndProviders();
+const App = () => (
+  <AppErrorBoundary>
+    <AppListener />
 
-  return (
-    <AppErrorBoundary>
-      <AppListener />
-
-      <FileManagerContextProvider>
-        <PersistedDirectoriesContextProvider>
-          {/* <Routes>
-              <Route index element={<Navigate to={PageRoute.STARTUP} />} />
-              {startupRoutes}
-              {animatorRoutes}
-            </Routes> */}
-          <Outlet />
-        </PersistedDirectoriesContextProvider>
-      </FileManagerContextProvider>
-    </AppErrorBoundary>
-  );
-};
+    <FileManagerContextProvider>
+      <PersistedDirectoriesContextProvider>
+        <Outlet />
+      </PersistedDirectoriesContextProvider>
+    </FileManagerContextProvider>
+  </AppErrorBoundary>
+);
 
 export default App;
